Get rid of webpack-merge
