+++ ui.apps/src/main/webpack.core/package.json
@@ -40,8 +40,7 @@
     "sass-loader": "^7.0.3",
     "stylelint": "^9.3.0",
     "webpack": "^4.15.1",
-    "webpack-command": "^0.4.0",
-    "webpack-merge": "^4.1.3"
+    "webpack-command": "^0.4.0"
   },
   "dependencies": {
     "babel-polyfill": "^6.26.0"
